Fix UBSan errors reported in the testsuite using gcc --version:
gcc (Ubuntu 7.4.0-1ubuntu1~18.04.1) 7.4.0

- Left-shift into sign-bit.
- Signed integer overflow when modulo behavior is intended
- Use memcpy (which modern compilers should optimize anyway) to avoid unaligned reads

+++ c++/src/RLE.hh
@@ -27,7 +27,8 @@
 namespace orc {
 
   inline int64_t zigZag(int64_t value) {
-    return (value << 1) ^ (value >> 63);
+    // avoid UB with left-shift casting to unsigned
+    return static_cast<int64_t>(static_cast<uint64_t>(value) << 1) ^ (value >> 63);
   }
 
   inline int64_t unZigZag(uint64_t value) {
